backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djoser==2.1.0
django-cors-headers==3.13.0
django-filter==2.4.0
drf-extra-fields==3.4.0
Pillow==10.0.0
psycopg2-binary==2.9.3
python-dotenv
django-colorfield
drf-base64
reportlab
fpdf